refactor: move bot logic to module
